pyproject.toml
[project]
name = "vrdj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Brett Viren", email = "brett.viren@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "beets>=2.5.1",
    "click>=8.3.1",
    "faiss-cpu>=1.13.0",
    "numpy>=2.3.5",
    "soundfile>=0.13.1",
    "torch>=2.9.1",
    "torchvggish>=0.2",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project.scripts]
vrdj = "vrdj.cli:main"
